# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/User/show": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "show a User",
                "operationId": "c64d3d5f769c326b2963b116fc7022e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id User",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User detail successfully"
                    }
                }
            }
        },
        "/api/User/destroy": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete a User",
                "operationId": "9cc85009dccf3f1f89e09f79c2431a0c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id User",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Delete successfully"
                    }
                }
            }
        },
        "/api/User/restore": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "restore a User",
                "operationId": "ad70f3b2d6965b237fc0b91c0711b3da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id User",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User restore successfully"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}